+++ IntersectAABB_shared.h
@@ -45,7 +45,7 @@ public:
 	AABBNODE *left;
 	AABBNODE *right;
 	std::vector<SimulationFacet*> facets;
-	size_t GetMemSize();
+	int GetMemSize();
 };
 
 AABBNODE *BuildAABBTree(const std::vector<SimulationFacet *> &facets, const int depth, int& maxDepth);
